Utveckling av webbapplikationer/Flask/flask projekt/requirements.txt
flask
werkzeug
os
pillow